Move properties into the parent component
Apply on SidePanel

- Set layout related property in parent composents
- Extract some props in parent component
